Added test for entities that end  's' and has_one
The reason for this test is that release 0.0.8 has a bug where if you an
entity with the name 'Address' and another entity 'has_one :address',
the relationship is not properly handled.

In other words, if you run this test for verson 0.0.8, it fails. On
master, it succeeds. So this test is simply to prevent regression.

I believe the fix to this was simply when the code updated the version
of ActiveSupport.